requirements.txt
mtcnn==0.1.0
numpy==1.19.5
streamlit==0.79.0
opencv_python_headless==4.5.1.48
Pillow==8.2.0
